+++ src/main/java/pl/asie/lib/chat/CommandNick.java
@@ -56,7 +56,7 @@ public class CommandNick extends CommandBase {
 			}
 
 			AsieLibMod.nick.setNickname(target, newName);
-			sender.addChatMessage(new ChatComponentTranslation("commands.nick.done", AsieLibMod.nick.getNickname(target)));
+			AsieLibMod.chat.sendChatMessage(sender, new ChatComponentTranslation("commands.nick.done", AsieLibMod.nick.getNickname(target)));
 		} else {
 			throw new WrongUsageException(this.getCommandUsage(sender));
 		}