+++ src/main/java/pl/asie/lib/chat/NicknameRepository.java
@@ -6,6 +6,7 @@ import com.google.gson.Gson;
 import com.google.gson.internal.LinkedTreeMap;
 import com.google.gson.reflect.TypeToken;
 import cpw.mods.fml.common.FMLCommonHandler;
+import cpw.mods.fml.common.eventhandler.EventPriority;
 import cpw.mods.fml.common.eventhandler.SubscribeEvent;
 import cpw.mods.fml.relauncher.Side;
 import net.minecraft.entity.player.EntityPlayer;
@@ -27,25 +28,28 @@ public class NicknameRepository implements INicknameRepository {
 	private Gson gson;
 	private Type gsonType;
 	private HashSet<INicknameHandler> handlers = new HashSet<INicknameHandler>();
-	
+
 	public NicknameRepository() {
 		gson = new Gson();
-		gsonType = new TypeToken<HashBiMap<String, String>>(){}.getType();
+		gsonType = new TypeToken<HashBiMap<String, String>>() {
+		}.getType();
 		nicknames = HashBiMap.create();
 	}
-	
+
 	public void setNickname(String username, String nickname) {
 		//AsieLibMod.log.info("Trying to set " + username + " to nickname " + nickname);
-		if(nickname.equals("-") && nicknames.containsKey(username)) {
-			nicknames.remove(username);
+		if(nickname.equals("-") || nickname.equals(username)) {
+			if(nicknames.containsKey(username)) {
+				nicknames.remove(username);
+			}
 		} else {
 			nicknames.put(username, nickname);
 		}
-		
+
 		updateNickname(username);
 		saveNicknames();
 	}
-	
+
 	public void updateNickname(String username) {
 		EntityPlayer player = PlayerUtils.find(username);
 		//AsieLibMod.log.info("Trying to update nickname " + username);
@@ -53,23 +57,32 @@ public class NicknameRepository implements INicknameRepository {
 			//AsieLibMod.log.info("Trying to update display name " + username);
 			player.refreshDisplayName();
 		}
-		if(FMLCommonHandler.instance().getEffectiveSide() != Side.SERVER) return;
+		if(FMLCommonHandler.instance().getEffectiveSide() != Side.SERVER) {
+			return;
+		}
 		//AsieLibMod.log.info("Updating nickames on server... ");
 		String nickname = getNickname(username);
-		for(INicknameHandler handler: handlers) {
+		for(INicknameHandler handler : handlers) {
 			handler.onNicknameUpdate(username, nickname);
 		}
 	}
-	
+
 	public String getUsername(String nickname) {
-		BiMap<String, String> usernames = nicknames.inverse();
-		return usernames.containsKey(nickname) ? usernames.get(nickname) : null;
+		//BiMap<String, String> usernames = nicknames.inverse();
+		for(String username : nicknames.keySet()) {
+			String nick = nicknames.get(username);
+			if(nick != null && nick.replaceAll("&.", "").equals(nickname)) {
+				return username;
+			}
+		}
+		//return usernames.containsKey(nickname) ? usernames.get(nickname) : null;
+		return null;
 	}
-	
+
 	public String getNickname(String username) {
 		return nicknames.containsKey(username) ? nicknames.get(username) : username;
 	}
-	
+
 	public String getRawNickname(String username) {
 		String nn = getNickname(username);
 		return nn.replaceAll("&.", "");
@@ -78,12 +91,14 @@ public class NicknameRepository implements INicknameRepository {
 	public boolean hasNickname(String username) {
 		return nicknames.containsKey(username) && !nicknames.get(username).equals(username);
 	}
-	
+
 	public void loadNicknames(String data) {
-		if(FMLCommonHandler.instance().getEffectiveSide() != Side.SERVER) return;
+		if(FMLCommonHandler.instance().getEffectiveSide() != Side.SERVER) {
+			return;
+		}
 		try {
 			nicknames = gson.fromJson(data, gsonType);
-			for(String username: nicknames.keySet()) {
+			for(String username : nicknames.keySet()) {
 				updateNickname(username);
 			}
 		} catch(Exception e) {
@@ -91,9 +106,11 @@ public class NicknameRepository implements INicknameRepository {
 			nicknames = HashBiMap.create();
 		}
 	}
-	
+
 	public void loadNicknames() {
-		if(FMLCommonHandler.instance().getEffectiveSide() != Side.SERVER) return;
+		if(FMLCommonHandler.instance().getEffectiveSide() != Side.SERVER) {
+			return;
+		}
 		nicknameFile = new File(DimensionManager.getCurrentSaveRootDirectory(), "nicknames.json");
 		nicknames = HashBiMap.create();
 		if(!nicknameFile.exists()) {
@@ -101,19 +118,21 @@ public class NicknameRepository implements INicknameRepository {
 		}
 		try {
 			LinkedTreeMap<String, String> hm = gson.fromJson(new FileReader(nicknameFile), gsonType);
-			for(String a: hm.keySet()) {
+			for(String a : hm.keySet()) {
 				nicknames.put(a, hm.get(a));
 			}
-			for(String username: nicknames.keySet()) {
+			for(String username : nicknames.keySet()) {
 				updateNickname(username);
 			}
 		} catch(Exception e) {
 			e.printStackTrace();
 		}
 	}
-	
+
 	public void saveNicknames() {
-		if(FMLCommonHandler.instance().getEffectiveSide() != Side.SERVER) return;
+		if(FMLCommonHandler.instance().getEffectiveSide() != Side.SERVER) {
+			return;
+		}
 		try {
 			FileWriter fw = new FileWriter(nicknameFile);
 			fw.write(toString());
@@ -122,12 +141,12 @@ public class NicknameRepository implements INicknameRepository {
 			e.printStackTrace();
 		}
 	}
-	
+
 	public String toString() {
 		return gson.toJson(nicknames, gsonType);
 	}
-	
-	@SubscribeEvent
+
+	@SubscribeEvent(priority = EventPriority.HIGH)
 	public void nameFormat(PlayerEvent.NameFormat event) {
 		event.displayname = getRawNickname(event.username);
 		//AsieLibMod.log.info("Name format change: Changed " +  event.username + " to " + event.displayname);
@@ -135,5 +154,9 @@ public class NicknameRepository implements INicknameRepository {
 
 	public void addHandler(INicknameHandler handler) {
 		handlers.add(handler);
+	}
+
+	public boolean isNicknamesNull() {
+		return nicknames == null;
 	}
 }